Updating pattern to define service interfaces within the handler for the required functions for the REST interface. This will mean that the controllers don't have access to every method, only the ones that are needed for the endpoints.
I came to this understanding because of an article explaining that golang interfaces are defined by consumers for their uses not by producers defining the list of all functions. I like the controller example because other services may need other functions but they wont be exposed to people writing controller endpoints because they aren't included in the purpose built interface.